Extract built-in identifier loader
