+++ src/plugin/src/project-index/index.js
@@ -1,6 +1,4 @@
 import path from "node:path";
-import { fileURLToPath } from "node:url";
-
 import { cloneLocation } from "../../../shared/ast-locations.js";
 import { getCallExpressionIdentifier } from "../../../shared/ast-node-helpers.js";
 import {
@@ -15,8 +13,7 @@ import {
 import {
     assertFunction,
     getOrCreateMapEntry,
-    hasOwn,
-    isPlainObject
+    hasOwn
 } from "../../../shared/object-utils.js";
 import {
     buildLocationKey,
@@ -29,12 +26,7 @@ import {
     isProjectManifestPath
 } from "./constants.js";
 import { defaultFsFacade } from "./fs-facade.js";
-import {
-    isFsErrorCode,
-    listDirectory,
-    getFileMtime
-} from "../../../shared/fs-utils.js";
-import { areNumbersApproximatelyEqual } from "../../../shared/number-utils.js";
+import { isFsErrorCode, listDirectory } from "../../../shared/fs-utils.js";
 import {
     getDefaultProjectIndexCacheMaxSize,
     loadProjectIndexCache,
@@ -48,12 +40,11 @@ import {
     createAbortGuard,
     throwIfAborted
 } from "../../../shared/abort-utils.js";
-import { parseJsonWithContext } from "../../../shared/json-utils.js";
-import { normalizeIdentifierMetadataEntries } from "../../../shared/identifier-metadata.js";
 import {
     analyseResourceFiles,
     createFileScopeDescriptor
 } from "./resource-analysis.js";
+import { loadBuiltInIdentifiers } from "./built-in-identifiers.js";
 
 const defaultProjectIndexParser = getDefaultProjectIndexParser();
 
@@ -289,107 +280,6 @@ export {
     PROJECT_INDEX_GML_CONCURRENCY_ENV_VAR,
     PROJECT_INDEX_GML_CONCURRENCY_BASELINE
 } from "./concurrency.js";
-
-const GML_IDENTIFIER_FILE_PATH = fileURLToPath(
-    new URL("../../../../resources/gml-identifiers.json", import.meta.url)
-);
-
-let cachedBuiltInIdentifiers = null;
-
-function extractBuiltInIdentifierNames(payload) {
-    if (!isPlainObject(payload)) {
-        throw new TypeError(
-            "Built-in identifier metadata must be an object payload."
-        );
-    }
-
-    const { identifiers } = payload;
-    if (!isPlainObject(identifiers)) {
-        throw new TypeError(
-            "Built-in identifier metadata must expose an identifiers object."
-        );
-    }
-
-    const entries = normalizeIdentifierMetadataEntries(payload);
-    const names = new Set();
-
-    for (const { name, type } of entries) {
-        if (type.length === 0) {
-            continue;
-        }
-
-        names.add(name);
-    }
-
-    return names;
-}
-
-function parseBuiltInIdentifierNames(rawContents) {
-    const payload = parseJsonWithContext(rawContents, {
-        source: GML_IDENTIFIER_FILE_PATH,
-        description: "built-in identifier metadata"
-    });
-
-    return extractBuiltInIdentifierNames(payload);
-}
-
-async function loadBuiltInIdentifiers(
-    fsFacade = defaultFsFacade,
-    metrics = null,
-    options = {}
-) {
-    const { signal, ensureNotAborted } = createAbortGuard(options, {
-        fallbackMessage: PROJECT_INDEX_BUILD_ABORT_MESSAGE
-    });
-
-    const currentMtime = await getFileMtime(
-        fsFacade,
-        GML_IDENTIFIER_FILE_PATH,
-        { signal }
-    );
-    ensureNotAborted();
-    const cached = cachedBuiltInIdentifiers;
-    const cachedMtime = cached?.metadata?.mtimeMs ?? null;
-
-    if (!cached) {
-        metrics?.recordCacheMiss("builtInIdentifiers");
-    } else if (
-        cachedMtime === currentMtime ||
-        (typeof cachedMtime === "number" &&
-            typeof currentMtime === "number" &&
-            areNumbersApproximatelyEqual(cachedMtime, currentMtime))
-    ) {
-        metrics?.recordCacheHit("builtInIdentifiers");
-        return cached;
-    } else {
-        metrics?.recordCacheStale("builtInIdentifiers");
-    }
-
-    let names = new Set();
-
-    try {
-        const rawContents = await fsFacade.readFile(
-            GML_IDENTIFIER_FILE_PATH,
-            "utf8"
-        );
-        ensureNotAborted();
-        names = parseBuiltInIdentifierNames(rawContents);
-    } catch {
-        // Built-in identifier metadata ships with the formatter bundle; if the
-        // file is missing or unreadable we intentionally degrade to an empty
-        // set rather than aborting project indexing. That keeps the CLI usable
-        // when installations are partially upgraded or when read permissions
-        // are restricted, and the metrics recorder above still notes the cache
-        // miss for observability.
-    }
-
-    cachedBuiltInIdentifiers = {
-        metadata: { mtimeMs: currentMtime },
-        names
-    };
-
-    return cachedBuiltInIdentifiers;
-}
 
 function createProjectTreeCollector(metrics = null) {
     const yyFiles = [];
@@ -2369,7 +2259,10 @@ export async function buildProjectIndex(
     });
 
     const builtInIdentifiers = await metrics.timeAsync("loadBuiltIns", () =>
-        loadBuiltInIdentifiers(fsFacade, metrics, { signal })
+        loadBuiltInIdentifiers(fsFacade, metrics, {
+            signal,
+            fallbackMessage: PROJECT_INDEX_BUILD_ABORT_MESSAGE
+        })
     );
     ensureNotAborted();
     const builtInNames = builtInIdentifiers.names ?? new Set();
@@ -2459,4 +2352,4 @@ export async function buildProjectIndex(
 }
 export { getDefaultFsFacade } from "./fs-facade.js";
 export { getProjectIndexParserOverride };
-export { loadBuiltInIdentifiers as __loadBuiltInIdentifiersForTests };
+export { __loadBuiltInIdentifiersForTests } from "./built-in-identifiers.js";
